Use ansi-term as the default shell again
Eshell does not work well with pipes.

If I `cat` a long file and pipe to `head -10`,
it is slow to output and the following error occurs:

```
error in process filter: eshell-output-object-to-target:
Writing to process: Input/output error, head
```

Emacs will use 100% of the CPU afterwards and needs
to be restarted.