Allow options/args transformations/shortcuts in helm-ag queries
This makes it possible to preprocess query arguments on the fly.

One use case (which is already implemented in this commit) is
the transformation of `-Gpattern` to `--iglob=**/*pattern*/**`,
which makes it possible to easily restrict rg searches to files
containing `pattern` in the path.

ag's -g flag probably will never be implemented in rg:
https://github.com/BurntSushi/ripgrep/issues/91.